+++ packages/xahau/src/models/ledger/index.ts
@@ -3,6 +3,7 @@ import AccountRoot, {
   AccountRootFlagsInterface,
 } from './AccountRoot'
 import Amendments, { Majority, AMENDMENTS_ID } from './Amendments'
+import AMM, { VoteSlot } from './AMM'
 import Check from './Check'
 import Cron from './Cron'
 import DepositPreauth from './DepositPreauth'
@@ -34,6 +35,8 @@ export {
   AccountRoot,
   AccountRootFlags,
   AccountRootFlagsInterface,
+  AMM,
+  VoteSlot,
   AMENDMENTS_ID,
   Amendments,
   Check,