+++ packages/xahau/HISTORY.md
@@ -34,6 +34,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Added
 * parseTransactionFlags as a utility function in the xrpl package to streamline transactions flags-to-map conversion
+* Support for the AMMClawback amendment (XLS-73)
 * Support for XLS-77d Deep-Freeze amendment
 
 ### Fixed